refactor: extract runtime minification plugin
